+++ src/gui/fill_stroke_widget.cpp
@@ -54,9 +54,9 @@ fill_stroke_widget::fill_stroke_widget (gui_model *model, QWidget *parent)
   swap_fill_stroke_button->setAutoRaise (true);
 
 
-  m_view->add_gui_widget (style_controller_role_t::FILL_SERVER, new color_indicator_gui_widget (m_fill));
-  m_view->add_gui_widget (style_controller_role_t::STROKE_SERVER, new color_indicator_gui_widget (m_stroke));
-  m_view->add_gui_widget (style_controller_role_t::IS_SELECTED_FILL, new fill_stroke_gui_widget (this, m_fill, m_stroke));
+  m_view->add_gui_widget ((int)style_controller_role_t::FILL_SERVER, new color_indicator_gui_widget (m_fill));
+  m_view->add_gui_widget ((int)style_controller_role_t::STROKE_SERVER, new color_indicator_gui_widget (m_stroke));
+  m_view->add_gui_widget ((int)style_controller_role_t::IS_SELECTED_FILL, new fill_stroke_gui_widget (this, m_fill, m_stroke));
 
   QSize size = m_fill->sizeHint ();
   m_fill->setGeometry (QRect (0, 0, size.width (), size.height ()));
@@ -64,7 +64,7 @@ fill_stroke_widget::fill_stroke_widget (gui_model *model, QWidget *parent)
   m_stroke->move (stroke_pos ());
   swap_fill_stroke_button->move (m_fill->sizeHint ().width (), 0);
   
-  set_is_selected_fill (m_model->data (style_controller_role_t::IS_SELECTED_FILL).toBool ());
+  set_is_selected_fill (m_model->data ((int)style_controller_role_t::IS_SELECTED_FILL).toBool ());
 
   CONNECT (swap_fill_stroke_button, &QToolButton::clicked, this, &fill_stroke_widget::swap_fill_stroke);
 }
@@ -97,10 +97,10 @@ void fill_stroke_widget::set_is_selected_fill (bool selected)
 void fill_stroke_widget::swap_fill_stroke ()
 {
   auto changer = m_model->do_multi_change ();
-  QVariant stroke = changer->data (style_controller_role_t::STROKE_SERVER);
-  QVariant fill = changer->data (style_controller_role_t::FILL_SERVER);
-  changer->set_data (style_controller_role_t::FILL_SERVER, stroke);
-  changer->set_data (style_controller_role_t::STROKE_SERVER, fill);
+  QVariant stroke = changer->data ((int)style_controller_role_t::STROKE_SERVER);
+  QVariant fill = changer->data ((int)style_controller_role_t::FILL_SERVER);
+  changer->set_data ((int)style_controller_role_t::FILL_SERVER, stroke);
+  changer->set_data ((int)style_controller_role_t::STROKE_SERVER, fill);
 }
